requirements.txt
numpy==1.19.1
tqdm==4.36.1
six==1.12.0
tensorboard==2.0.0
matplotlib==3.4.3
pandas==1.3.2
Pillow==8.3.1
scikit_learn==0.24.2
scipy==1.7.1
timm==0.4.12
torch==1.9.0
torchvision==0.10.0
